adjusts migration on accounts to be an integer. formats responses

+++ db/migrations/001_create_accounts.rb
@@ -2,7 +2,7 @@ Sequel.migration do
   change do
     create_table(:accounts) do
       primary_key :id
-      Float :balance, :null => false
+      Integer :balance, :null => false
     end
   end
 end
